+++ data-raw/full_bma_results.R
@@ -12,6 +12,11 @@ data_prepared <- badp::economic_growth %>%
     scale         = FALSE
   )
 
-full_bma_results <- badp::bma(full_model_space, round = 5)
+# Note: the robust standard deviations (PSDR, PSDRcon) in this object are not
+# identified, because every model in full_model_space has more parameters than
+# there are entities. See the "Rank of the sandwich covariance" section of
+# ?optim_model_space. They are retained because bma() reports them, but they
+# should not be used as a reference.
+full_bma_results <- badp::bma(badp::full_model_space, round = 5)
 
 usethis::use_data(full_bma_results, overwrite = TRUE)